Feature #11551
Fix the code and the tests with the upgrade of some dependencies defined in the
top level definition of the Silverpeas projects (bom of dependencies).

For example, a bug has been fixed in iCal4J: when a recurrent event is
defined from a given date and for some days of week that doesn't match the
start date of the event, then now no occurrence of the event is occurring at
this start date as it doesn't match the recurrence rule.

Another example with Hibernate implementation of JPA: the getting of an entity
fails when that object requires (non-null dependency) another entity that has
been deleted, even with a lazy fetching strategy. Refine our own JPA base model
to be more compliant to Hibernate constraints on its own implementation of JPA.

Now Silverpeas is based upon JEE 8.

Specify the Docker container to use to execute the build with Java 11 and
Wildfly 20.

Add some unit tests.

In Jenkinsfile, compute the version of the parent pom when a snapshot version
of it is declared in the project's pom.

+++ core-library/src/main/java/org/silverpeas/core/contribution/converter/openoffice/OpenOfficeService.java
@@ -24,8 +24,8 @@
 
 package org.silverpeas.core.contribution.converter.openoffice;
 
-import org.jodconverter.office.LocalOfficeManager;
-import org.jodconverter.office.OfficeManager;
+import org.jodconverter.local.office.LocalOfficeManager;
+import org.jodconverter.core.office.OfficeManager;
 import org.silverpeas.core.initialization.Initialization;
 import org.silverpeas.core.util.SettingBundle;
 import org.silverpeas.core.util.StringUtil;
@@ -53,7 +53,7 @@ public class OpenOfficeService implements Initialization {
   public void init() throws Exception {
     final String home = settings.getString(OPENOFFICE_HOME);
     final String ports = settings.getString(OPENOFFICE_PORT, "8100");
-    final int taskTimeout = settings.getInteger(OPENOFFICE_TASK_TIMEOUT, 30000);
+    final long taskTimeout = settings.getLong(OPENOFFICE_TASK_TIMEOUT, 30000);
     final int[] portNumbers = Stream.of(ports.split(","))
         .map(String::trim)
         .mapToInt(Integer::parseInt)
